Changed: Do not check preexec and chpwd hooks
Turns out that when using git aliases, e.g gcheckout -b foo, the command
prompt won't be updated unless oh-my-zsh is active. The check is
supposed to compare the current command in the history and see if it
contains *git* or *vcs*. The code sets a flag which then conditionally
updates the prompt like so:

case "$(history $HISTCMD)" in
    *git*)
        PR_GIT_UPDATE=1
        ;;
    *svn*)
        PR_GIT_UPDATE=1
        ;;
esac

The problem is that when it retrieves the last command and compares it
to *git* and that command is a git alias (gcheckout) the check will
fail. The reason it succeeds when oh-my-zsh is enabled is because
oh-my-zsh aliases `history` to `fc -l 1`. This means that the entire
history is searched, and if you happen to have a git command in there
the prompt will update.

I don't understand why this check was introduced as it seems like a
premature optimization. So let's remove it.

Also changed the style of the prompt slightly.